refactor: public 폴더에서 이미지 관리
