+++ apps/web/public/images/index.ts
