+++ apps/web/public/images/logo-white.png
